Move assert in cmdServerSigChild().
It is possible for the server to receive notifications for children that were not spawned by it, e.g. when running as the root process of a container.

Only assert when the child process is found in our process list. Otherwise, the code may be something unexpected.

Since this is an assert there should not be any real world consequences either way, but the issue has caused problems in testing.