Disable arm64 test in CirrusCI.
There seems to be a shortage of arm64 hosts because queue times have been steadily increasing over the last few weeks. It can now take several hours to get an arm64 test queued, which makes it difficult to get development done.

Disable for the time being and hope the resource issue gets resolved in the future.